[FLINK-39537][table] Address feedback
